backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-core
supabase==1.0.4
pypdf
sentence-transformers
python-multipart